Added CustomCommands
- Added the functionality for custom commands (CustomCommandsService.cs), (CustomCommands.cs) and edited Program.cs.
- Added DSharpPlus.Interactivity NuGet package for the interactivity (waiting for message)

Custom Commands: The so-called CC's are a way to add commands to the bot without touching a single line of code. You can add a command and the content the bot is supposed to display once you invoke the command, through a simple Discord bot command (`newcc` with the (many) different aliases). Then you can specify the content, which is supposed to be displayed by the bot, once you invoke the command. So CC's will be purely informational commands, without any special code, that needs to be executed. The CC's and the content are saved in Text files.

The functionality is not done yet!